Create fetch history lists hook
