+++ src/global-state/shopListHistoryState.tsx
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const shopListHistoryState = atom({
+    key: 'shopListHistoryState',
+    default: [],
+})
